Support input params
