+++ .npmignore
@@ -240,7 +240,8 @@ markdowns/
 .babelrc
 # typings/
 # eslint-rules/
-scripts/
+scripts/*
+!scripts/prReleaseNotesCommon.ts
 demo-app.component.js
 index.android.js
 index.ios.js
